requirements-dev.txt
-r requirements.txt
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-aiohttp==1.1.0

